Lab1/requirements.txt
astroid==2.1.0
cycler==0.10.0
graphviz==0.10.1
isort==4.3.4
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
matplotlib==3.0.2
mccabe==0.6.1
numpy==1.15.4
progressbar2==3.38.0
pylint==2.2.2
pyparsing==2.3.0
python-dateutil==2.7.5
python-utils==2.3.0
scikit-learn==0.20.1
scipy==1.1.0
six==1.11.0
typed-ast==1.1.0
wrapt==1.10.11
